Use helpers, add id validation.
